Data Structure Chapter 1 I Arrays and Strings 1.7 Rotate Matrix
This is an 2nd attempt at the question, as a mean of revision and practice coding on paper.
commit add982 Copy straight from the code written on paper.
commit 12f942 Add in main method and helper method to test code. Fix bugs found on 2 lines and rotate method, which is outside the coding scope
commit 1ca621 Port the beautified code with comment to RotateMatrix class, update test case.